requirements.txt
flask==2.3.2
gunicorn==21.2.0
tensorflow-cpu==2.10.0
numpy==1.23.5
pillow
pandas
werkzeug==2.3.6
